backend/pyproject.toml
[project]
name = "edos-security-backend"
version = "1.0.0"
description = "EDoS Security Dashboard Backend - FastAPI backend for real-time cybersecurity monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "faker>=38.2.0",
    "fastapi>=0.121.3",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
